bounds.h: - changed num_bound_boxes from short to unsigned int
entity.h:
- removed #defines for MAX_ENT and ID_ENEMY (moved to enums.h)

enums.h:
- moved MAX_ENT and ID_ENEMY to here from entity.h
- added WINDOW_TILES_W/H for map editor

kq.h:
- removed commented-out code for portrait[MAXCHRS]

mainwindow.c:
- added extra spacing between functions
- moved declarations to beginning of functions (more C-style)
- added missing return value to on_map_button_release_event(), on_map_motion_notify_event()

mapdiff.c:
- create compiler error if Allegro predates 4.1, or is 4.9 or later (5.0 candidate)
- added extra spaces between functions

mapdraw2.h:
- changed max_sets from short to unsigned int

mapdump.c:
- create compiler error if Allegro predates 4.1, or is 4.9 or later (5.0 candidate)

mapedit.c:
- updated process_menu_bottom() for WarpX/Y displacement

mapshared.c:
- separated loops in visual_map() depending on which layer(s) were showing
- transitioning over to TILE_W and TILE_H

other/kq2h.pl:
- small formatting changes (if we ever use this to regenerate res.c...)

scripts/init.lua:
- cosmetic spacing

disk.c, effects.c, eqpmenu.c, intrface.c, magic.c:
- transitioning to preprocessor defines

draw.c:
- optimized drawing routine within draw_playerbound() and draw_char()

entity.c:
- cosmetic spacing
- transitioning to preprocessor defines

prepare_map():
- unrolled loops

+++ KQ/include/bounds.h
@@ -60,7 +60,7 @@ int bound_in_bound2 (s_bound *, s_bound *, int);
 s_bound *is_contained_bound (s_bound *, unsigned int, int, int, int, int);
 void set_bounds (s_bound *, int, int, int, int);
 
-extern short num_bound_boxes;
+extern unsigned int num_bound_boxes;
 extern s_bound bound_box[MAX_BOUNDS];
